towards downloading and opening zipped csv files. just have to figure out 'unknown escape character' error and we should be good to go
